api/pyproject.toml
[project]
name = "digital-closet-api"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "chrisyecora",email = "chris.yecora@gmail.com"}
]
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "boto3 (>=1.42.72,<2.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pgvector (>=0.3.6,<0.4.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "email-validator (>=2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "pillow-heif (>=1.3.0,<2.0.0)",
    "torch (==2.2.2)",
    "torchvision (==0.17.2)",
    "ultralytics (>=8.4.26,<9.0.0)",
    "transformers (<4.40.0)",
    "httpx (>=0.20.0)",
    "pytest-httpx (>=0.36.0,<0.37.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "numpy (<2)"
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
